+++ src/main/java/uk/ac/ebi/eva/pipeline/jobs/steps/GenesLoad.java
@@ -29,7 +29,7 @@ import org.springframework.context.annotation.Import;
 import uk.ac.ebi.eva.pipeline.configuration.VariantJobsArgs;
 import uk.ac.ebi.eva.pipeline.io.readers.GeneReader;
 import uk.ac.ebi.eva.pipeline.io.writers.GeneWriter;
-import uk.ac.ebi.eva.pipeline.jobs.processors.GeneFilterProcessor;
+import uk.ac.ebi.eva.pipeline.jobs.steps.processors.GeneFilterProcessor;
 import uk.ac.ebi.eva.pipeline.listeners.SkipCheckingListener;
 import uk.ac.ebi.eva.pipeline.model.FeatureCoordinates;
