+++ src/main/java/uk/ac/ebi/eva/pipeline/jobs/steps/VariantsAnnotGenerateInput.java
@@ -20,7 +20,7 @@ import com.mongodb.DBObject;
 import uk.ac.ebi.eva.pipeline.configuration.VariantJobsArgs;
 import uk.ac.ebi.eva.pipeline.io.readers.VariantReader;
 import uk.ac.ebi.eva.pipeline.io.writers.VepInputWriter;
-import uk.ac.ebi.eva.pipeline.jobs.processors.VariantAnnotationItemProcessor;
+import uk.ac.ebi.eva.pipeline.jobs.steps.processors.VariantAnnotationItemProcessor;
 import uk.ac.ebi.eva.pipeline.model.VariantWrapper;
 
 import org.slf4j.Logger;
