+++ src/main/java/uk/ac/ebi/eva/pipeline/jobs/steps/processors/GeneFilterProcessor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.pipeline.jobs.processors;
+package uk.ac.ebi.eva.pipeline.jobs.steps.processors;
 
 import org.springframework.batch.item.ItemProcessor;
